# apps/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MX Space Core API",
    "version": "v3",
    "description": "Endpoints consumed by the Space iOS client. Generated from Zod schemas — do not edit by hand."
  },
  "servers": [
    {
      "url": "/api/v3",
      "description": "Self-hosted instance, relative to the configured host"
    }
  ],
...
  "paths": {
...
    "/activity/recent": {
      "get": {
        "operationId": "getRecentActivities",
        "summary": "Recent likes and comments across published content",
        "tags": [
          "movement"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RecentActivities"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/recently": {
      "get": {
        "operationId": "listRecently",
        "summary": "Cursor-paged recently entries with hydrated media enrichments",
        "tags": [
          "recently"
        ],
        "parameters": [
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RecentlyCard"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createRecently",
        "summary": "Post a recently entry",
        "tags": [
          "recently"
        ],
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecentlyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RecentlyDetail"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/recently/ref-candidates": {
      "get": {
        "operationId": "listRecentlyRefCandidates",
        "summary": "Search internal content that can be attached as Recently context",
        "tags": [
          "recently"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "default": 12,
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RecentlyRefCandidate"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/recently/{id}": {
      "put": {
        "operationId": "updateRecently",
        "summary": "Rewrite a recently entry",
        "tags": [
          "recently"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecentlyCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RecentlyDetail"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteRecently",
        "summary": "Delete a recently entry",
        "tags": [
          "recently"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "5XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResponseMeta": {
        "type": "object",
        "additionalProperties": true
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "RecentActivities": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "author": {
                  "type": "string"
                },
                "avatar": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "id": {
                  "type": "string"
                },
                "nid": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "slug": {
                  "type": "string"
                },
                "text": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              },
              "required": [
                "author",
                "created_at",
                "text"
              ],
              "additionalProperties": false
            }
          },
          "like": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "id": {
                  "type": "string"
                },
                "nid": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "slug": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              },
              "required": [
                "created_at",
                "id"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "comment",
          "like"
        ],
        "additionalProperties": {}
      },
...
      "EnrichmentResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subtype": {
            "type": "string"
          },
          "published_at": {
            "type": "string"
          },
          "fetched_at": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "thumbnail_image": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string"
              },
              "width": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "height": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "alt": {
                "type": "string"
              },
              "thumbhash": {
                "type": "string"
              },
              "palette": {
                "type": "object",
                "properties": {
                  "dominant": {
                    "type": "string"
                  },
                  "swatches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "dominant"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "url"
            ],
            "additionalProperties": {}
          },
          "preview_image": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string"
              },
              "width": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "height": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "alt": {
                "type": "string"
              },
              "thumbhash": {
                "type": "string"
              },
              "palette": {
                "type": "object",
                "properties": {
                  "dominant": {
                    "type": "string"
                  },
                  "swatches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "dominant"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "url"
            ],
            "additionalProperties": {}
          },
          "capture_image": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string"
              },
              "width": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "height": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "alt": {
                "type": "string"
              },
              "thumbhash": {
                "type": "string"
              },
              "palette": {
                "type": "object",
                "properties": {
                  "dominant": {
                    "type": "string"
                  },
                  "swatches": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "dominant"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "url"
            ],
            "additionalProperties": {}
          },
          "attributes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "value": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    }
                  ]
                },
                "label": {
                  "type": "string"
                },
                "format": {
                  "type": "string",
                  "enum": [
                    "number",
                    "rating",
                    "date",
                    "percent",
                    "text",
                    "duration"
                  ]
                }
              },
              "required": [
                "key",
                "value"
              ],
              "additionalProperties": {}
            }
          },
          "links": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rel": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "rel",
                "url"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "title",
          "url",
          "category",
          "fetched_at"
        ],
        "additionalProperties": {}
      },
      "RecentlyCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time"
          },
          "ref_type": {
            "type": "string"
          },
          "ref_id": {
            "type": "string"
          },
          "ref": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "post",
                  "note",
                  "page",
                  "recently"
                ]
              },
              "title": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "nid": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "url": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "type"
            ],
            "additionalProperties": false
          },
          "up": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "down": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "comments_index": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "allow_comment": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "selected_enrichment_urls": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "additionalProperties": {}
          },
          "enrichments": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/EnrichmentResult"
            }
          }
        },
        "required": [
          "id",
          "content",
          "type",
          "created_at"
        ],
        "additionalProperties": {}
      },
      "RecentlyRefCandidate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "post",
              "note",
              "page",
              "recently"
            ]
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type"
        ],
        "additionalProperties": false
      },
      "RecentlyCreate": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1
          },
          "ref": {
            "type": "string"
          },
          "refType": {
            "type": "string",
            "enum": [
              "post",
              "note",
              "page",
              "recently"
            ]
          },
          "clearRef": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "selectedEnrichmentUrls": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "content"
        ]
      },
      "RecentlyDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time"
          },
          "ref_type": {
            "type": "string"
          },
          "ref_id": {
            "type": "string"
          },
          "ref": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "post",
                  "note",
                  "page",
                  "recently"
                ]
              },
              "title": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "nid": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "url": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "type"
            ],
            "additionalProperties": false
          },
          "up": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "down": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "comments_index": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "allow_comment": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "selected_enrichment_urls": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            },
            "additionalProperties": {}
          },
          "enrichments": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/EnrichmentResult"
            }
          }
        },
        "required": [
          "id",
          "content",
          "type",
          "created_at"
        ],
        "additionalProperties": {}
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}